split index to multiple files
