+++ src/helper.ts
@@ -1,0 +1,11 @@
+export const labelMap: Array<Array<number>> = [
+  //0  1  2  3  4  5  6  7  8  9 10 11   // align flags
+  [0, 6, 2, 8, 9, 11, 3, 5, 1, 4, 7, 10], // 0 = no centering
+  [1, 7, -1, -1, 9, 11, 4, -1, -1, -1, -1, 10], // 1 = center x
+  [3, -1, 5, -1, 9, 11, -1, -1, 4, -1, -1, 10], // 2 = center y
+  [4, -1, -1, -1, 9, 11, -1, -1, -1, -1, -1, 10], // 3 = center x & y
+  [0, 6, 2, 8, 10, -1, 3, 5, 1, 4, 7, -1], // 4 = center front (default)
+  [1, 7, -1, -1, 10, -1, 4, -1, -1, -1, -1, -1], // 5 = center front & x
+  [3, -1, 5, -1, 10, -1, -1, -1, 4, -1, -1, -1], // 6 = center front & y
+  [4, -1, -1, -1, 10, -1, -1, -1, -1, -1, -1, -1] // 7 = center front & x & y
+];
